Trimming attributes of Debt.
savings_rate and living_expenses have been removed,
along with max_payment and payment_from_savings.

If there's a desire to contribute to Debt accounts
(or any other accounts, for that matter)
from different pots of money, that should be handled by client code.

Corresponding changes made to tests for DebtPaymentStrategy.
(No changes required to DebtPaymentStrategy itself,
which is already a wrapper for TransactionStrategy).