Re-instating previous TransactionStrategy tests.
These are the same as the old tests, except that `total` is replaced
by `available` in  TransactionStrategy's call signature.

Also removed tests of init functionality.
Someone commented to me that if you have
a test for init, your init is too complicated.
But this init is super-basic, so why bother with specific tests?