Moving group/threshold logic to traversal class.
In the course of refactoring this code, we also added the `skip_nodes`
argument (which removes the need for the `reduce_node` method) and
simplified `_process_node_weighted` (in part by adding
the new `_reweight_children`, which can be accessed by threshold methods
to maintain consistency).

Determining weights for children now involves compensating for
transactions assigned during earlier
traversals. The current approach works (it's mathematically sound) but
the code is not at all self-explanatory.
Consider reimplementing by, say, prioritizing under-contributed children
(which would make it easy to determine thresholds too!)