Moving method to function.
The _underweighted_children method is now the _rebalance_children func.

This function is a work in progress; it is not called by any other code.
It might never be used.
C.f. issue #66, which proposes a different approach.

Also adds a kwarg argument to __call__.
This makes it easier to use TransactionTraversal as a dop-in replacement
for TransactionStrategy (which some tests do, and which
is recommended for programmatic interfaces)