Initial working graph-based traversal.
This version is quite limited.
It doesn't allocate minimum flows, only maximums.
It doesn't deal with per-node limits.
It throws an error if the max. flows could not be assigned
(because it throws an error on recursion generally).

But still, it's promising! It deals correctly with linked accounts,
which the old tree-based code had real trouble with.