Redesigning to use weighted edges, overflow nodes.
This removes the (recently-added) skip_nodes method,
along with the _reduce_graph method.

Instead of building the graph, iterating over it to assign flows,
modifying it as appropriaate, and reattempting,
this new approach involves iteratively testing flows during the
initial build, setting capacities and weights accordingly, and
then executing the max-flow just once to assign the "real" flows.

This implementation is incomplete. There need to be changes to
_build_graph to perform the iterative capacity-setting.
But methods for adding ordered and weighted nodes have
been updated as necessary, so this should be getting close!