Using flows to inform graph-building.
Before this change, we would build a graph by
traversing each node once, assign flows to edges,
then restrict the graph and repeat,
summing up the flows of each iteration to get the final flows.

The problem was that weighted nodes would often send less flow to
children than would be optimal ("underflows"), which threw off the
amount of flow assigned at various iterations.

Now we build the graph iteratively and assign flows once.
We build the graph, look for underflows,
restrict capacity accordingly, and then
generate flows once based on the restricted graph.

This has similar big-O time complexity to the earlier method
(and might even be slower on a variety of reasonable inputs),
since looking for underflows involves generating flows for the graph.
But it has the advantage of moving us towards correctness.