Adding `children` arg to `_add_node_*`
This commit removes the need to pass a separate `outbound_node` arg
to `_add_node_*` methods. Instead, each of those
methods is simplified to receive both `node` and `children` args,
with the proviso that `node` does not necessarily
refer to an object with any semantically-useful attributes.
(It's used only as an origin for edges to children.)

This commit uses this feature to complete an initial (untested)
implementation of `_restrict_node_underflows`.