Adding _add_edges_from method.
This method is trying to solve a problem that probably doesn't need
to be solved. It ensures that a bunch of edges, if added at once,
will collectively have a int-valued capacity equal to the rounded
sum of capacities of the inputs.

Realistically, we should just add a test to
_restrict_node_underflows to ensure that it doesn't recurse on
the first branch (matching total flow to outbound capacity).