Correcting behaviour - all tests pass!
This commit adds a number of convenience methods,
renames _restrict_node_underflows to _balance_flows,
and implements some important algorithmic changes.

The most important of these is modifying the capacity of
edges from a node's corresponding overflow_node to its children
to ensure that any new flow through overflow_node will
be allocated to the children being recursed on (and not to other
children, whose flows should remain static.)

The other important changes are redistributing _flow_
between children, not their existing capacity,
and ensuring that every step of recursion reduces the number of
children (except when reducing total capacity).

This file is now very long; many convenience methods
should likely be moved to a util.py file.
But for now: rejoice! It works!!!