Avoids negative capacity in _add_weighted_children
Previously, _add_weighted_children would simply generate
capacities for each child and reduce them by any memoized flows.
The problem is that this could result in negative capacity,
which is a problem in general (it's not supported by networkx),
but also in particular it requires us to re-weight all other children.
This wasn't happening, so some output could be incorrect.

This code tweaks _add_weighted_children to be recursive.
If negative capacities are detected, it deals with that (rounds up to 0)
and then recurses on the remaining children.