Implementing type annotations.
Added forecaster.typing to define various common types.

Refined the Money protocol to be generic. This deals with many errors.

Got started on accounts.base. Broken; still lots of work to do.

Next challenge: Defining MoneyFactory to be compatible with `float`.
(We probably need to remove `Money` from the signature's inupts -
restrict MoneyConvertible to allow numeric types only?
But then do we need to wrap any call to money_factory in an
 isinstance(x, Money) test?
 And if so, how do we get access to Money at runtime - another init arg?
 Need to chew on this some more.

Relates to #37 and #56.