Reading settings and constants from JSON files.
Relates to #12 and #80.

This change will break any code that imports from `canada.constants`.

Code that imports from `settings` or `canada.settings` should work.
(But note that the constant `canada.constants.CONTRIBUTION_WEIGHTS`
was misnamed; it should have been `saving_weights`. That bug is
corrected here. It appears that no code actually imported the
incorrect name, so this fix should be non-breaking.)

TODO: Tests need to be added for `Settings` and `SettingsCanada`.