Updating Strategy tests with MethodRegister logic.
Registered methods are now accessed by `.registered_methods`
instead of `.strategies`, invalid strategies raise a `KeyError` on
call rather than a `ValueError` on init, and there is no longer any
conversion of method keys on init - it's all done at call time.

None of these are significant changes in behaviour.
The goal is to harmonize and simplify the implementation and behaviour
of classes with registered methods, which this does.