Initial, testable ScenarioSampler implementation.
This almost certainly is broken in surprising ways, as it is
not yet untested. But it provides a basic design to build off of and
includes a first implementation of a sampling method.

Stubs for various sampling methods, as well as comments
on potential implementation approaches, are also provided.

Relates to #75.