Adding walk-forward and random sampling methods.
Removing stub for k-fold cross-validation, since we don't need to
reserve a validation set.

This is not yet tested. It's becoming clearer that it will
make more sense to subclass `ScenarioSampler` rather than
use on class and select samplers using `MethodRegister`.
So the next commit will probably refactor this code significantly.

Relates to #75.