Initial implementation of ConstantScenarioSampler.
This code doesn't represent covariance as richly as it could.
Right now it looks only at date ranges with data for all variables.
But it would be better to calculate the pairwise covariance for
each variable and build a covariance matrix from those,
since data can be (and frequently is) provided in different
time intervals for each variable.
Expanding this code to work with pairwise covariances will
be the focus of the next commit.

None of this is tested and will probably break horribly.

Relates to #75.