Calculating covariance with ddof=0.
`numpy.cov` defaults to `ddof=1`, unlike most other
numpy methods (including `numpy.var`). This lead to surprising results.
To ensure that the results computed by `MultivariateSampler`
align with what would be computed by `numpy.var`, all calls to
`numpy.cov` now provide the kwarg `ddof=0`.
See here for more:
https://stackoverflow.com/questions/21030668/why-do-numpy-cov-diagonal-elements-and-var-functions-have-different-values

This commit also adds a test for `MultivariateSampler.sample`
(which passes) and
a basic test of `WalkForwardSampler.sample` (which fails).

The test set for `MultivariateSampler` is complete... for now.
We can move on to `WalkForwardSampler`, which doubtless has its own
issues.

Relates to #75.