Implementing `WalkForwardSampler` init arg logic
The affected args are `synchronize`, `wrap_data`, and `interval`.

Default behaviour for `interval` is now to generate a walk-forward
sample of values without interpreting dates, returns, etc.
If `interval` is provided, returns are recalculated - but this is now
delayed until after start dates have been chosen to increase the number
of possible samples that can be drawn.

`wrap_data` and `synchronize` are now implemented.

Extensive changes to `util.regularize_returns`
(which was probably broken to begin with - it's not yet tested)
to support these changes to `WalkForwardSampler`.

Tests added for `WalkForwardSampler`. They pass,
but test coverage is incomplete - still need to write tests for
`wrap_data`, `interval`, and `num_samples`.

Relates to #75.