Fixing error raised in `regularize_returns`.
`regularize_returns` was previously setting `date` to be the first
suitable date within the range covered by the data, but then would
attempt to interpolate a return from an _earlier_ date, resulting in
an error. The solution is to push the date back, of course.
But this causes a new problem; because we're essentially treating the
period before the first date as lying outside the dataset (despite
the returns in the first date describing this period),
we're reducing the size of the output dataset by one or more.

So the fixed code avoids an error but results in data-loss. To avoid
that,
we need to support inspecting periods prior to the start date, which
will involve inferring the interval that the start date covers.
Fortunately, we already have code that does that:
`values_from_returns`.
This code should be refactored to use that function to expand the
scope of the dataset, which can then be converted back to returns.
(Bonus: converting values to returns is far easier to follow
than accumulating returns over an arbitrary interval.)

We should probably improve the inference of the pre-start-date interval
by `values_from_returns`. That function includes comments with
suggestions on how to do that.

Relates to #75.