Revising `regularize_returns` to limit data loss.
This code now correctly generates dates, which is a big step forward.
However, it doesn't generate correct returns for those dates.
That issue can be traced to
`accumulate_return`.
It seems like that function needs to get a lot more complicated to work.

This commit does not make use of `values_from_returns`.
That is probably a far less complex way forward, and should
be the focus of the next commit.

Relates to #75.