Fixing `util.accumulate_return` for easy inputs.
Part of this is correcting the math (needed to add 1 to returns when
accumulating and subtract 1 at the end)
and part of this is fixing basic logic errors
(like including the returns at the start of an interval rather than
the end - oops!)

`accumulate_return` now returns correct values when applied to
sequences of returns which include `start_date` and `end_date` as keys.
But it probably does not work correctly when either endpoint is
not a key in the sequence of returns. Need to add a test for this.
Part of the complexity is that `interpolate_returns` is probably broken,
and part of it is that we want to avoid double-counting.
(e.g. if `start_date` falls between keys, we can interpolate a value
for `start_date` based on the first key following `start_date` - but
then we need to discard that key to avoid double-counting).

Although we're getting closer to a robust implementation of
`regularize_returns`, it's still not clear whether this is a better
approach than simply converting to/from portfolio values...

While working on this, I had some thoughts about removing `lookahead`
args while allowing client code to obtain this functionality via
reversed inputs. I'd added a TODO on this subject.

Relates to #75.