Extending `WalkForwardSampler._get_valid_dates`.
The method now adds the first possible valid date, even if
it's not a key in the sequence of returns. This ensures that the
longest possible walk-forward sequence is always available to be
sampled. (This avoids rare errors that arise when the time period of the
sequence is about the same as the time period of the dataset and
`interval` doesn't line up with the intervals between dates in the
dataset.)

Relates to #75.