Adding `HistoricalValueReaderArray`
This class is broken due to a mismatch in how different methods
expect the `data` attr to be structured. Some treat it asa sequence
of 2-tuples, others treat it as a
sequence of columns from the CSV file.
We need to be consistent; we'll likely settle on the former for
consistency with the dict-based implementation.

Relates to #75.