Generating covariances based on aligned data.
This is how `_generate_coveriances` used to work, but a previous commit
simplified most of its logic to a single call to `numpy.cov`.
However, that will cause trouble with misaligned datasets.
(We need to add a test for such datasets!)
So now we're back to determining the pairwise covariance
for each pair of columns.

Extensive supported changes are made to `_align_data`,
which was out of date following its removal from `_generate_covariances`
in a prior commit.

Relates to #75.