Adding support for Monte Carlo simulations.
This adds a `sample` method which generates `Scenario` objects
via a `ScenarioSampler` (or takes them from a user-provided iterable)
and runs `run_forecast` over them.

`run_forecast` is extended to allow scenarios to be passed in as args.

It occurs to me that it may make sense to make `num_samples`
optional in `Forecaster`, `ScenarioSampler`, and samplers from
the `samplers` module. Will
look into this for a future commit.

This code is not yet tested.