Refactoring `ScenarioSampler` init args.
`num_samples` is now an optional kwarg with the same
semantics as `sampler` classes.
`data` is brought forward as the second positional arg.
This makes init calls much more readable.

`sampler_*` methods are now simplified, with `Scenario`-building
logic and `yield` statements moved to `__iter__`.

A new convenience method, `sample`, is now added. It simply
wraps `__iter__`.

Tests are updated to use the new call signature. All tests pass.

Note that I considered using different semantics for `num_samples`
than was used in the `sampler` module. In particular, I considered
causing `num_samples=None` to cause the _maximum_ number
of samples to be generated.
This could cause an endless loop if not used carefully, so
(following the principle of
least astonishment)
it now causes only one sample to be generated; if the user wants more,
it has to specify how many more.

Relates to #75.