Refactoring `Forecaster.sample`
`Forecaster.sample` now uses the refactored call signatures of
`ScenarioSampler` and has simplified some of its internal
naming and logic.

The only substantive change is a fix to a potential error (an exception
could be raised if passing a non-`ScenarioSampler` iterable `sampler`
arg with `num_samples=None`.
This is now fixed.)

Relates to #75.