Supporting replacing objects in function closures.
`forecaster.deepcopy`, in combination with `_populate_deepcopy_memo`,
will generate copies of objects where certain objects are replaced,
even when those objects are referenced in function closures.
This involves extending `copy.deepcopy`, and is a bit naughty
as it involves manipulating the `memo` arg to that function.

Adds a test of `deepcopy`.

Closes #82.