Monkey-patching `copy.deepcopy` to copy functions.
The previous implementation attempted to find functions
in the object reference graph independently and replace them
alongside calls to `copy.deepcopy`. This had some weaknesses;
references can be stored in various ways, which most of the
logic of `copy.deepcopy` is devoted to dealing with on a
case-by-case basis. It doesn't make sense to reimplement that
in parallel with `copy.deepcopy`; we need to extend `deepcopy` itself.

There are a few ways to do that. This one is the most dangerous,
as it is very implementation-specific, but it's also by far
the easiest, shortest, and most elegant. What is done here is
a new copier (the `copy` module's name for a function
that copies an object of a given type) is defined for
`FunctionType` and added to the proctected dict
`_deepcopy_dispatch`. (Accessing a protected name is the naughty part.)

Other approaches that might be worth considering: wrapping
`copy.deepcopy` and monkey-patching the wrapped
version back in to the `copy` module.
This is less naughty in that we would only be accessing public
names, but uglier in that we would need add a pre-processing
step to screen inputs for `FunctionType` and perhaps `memo` lookups
(which `deepcopy` handles already).

An even uglier but not at all naughty approach would be to
fork the implementation of `deepcopy` and make these changes
directly to the source. That might be worth doing if this
implementation breaks under a future version of `copy`.

All tests pass.

Relates to #82.